refactor: BaseEntity를 global 패키지로 이동
