+++ src/main/kotlin/com/nexters/bottles/global/BaseEntity.kt
@@ -1,4 +1,4 @@
-package com.nexters.bottles.user.domain
+package com.nexters.bottles.global
 
 import org.hibernate.annotations.CreationTimestamp
 import org.hibernate.annotations.UpdateTimestamp
